backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0
schedule==1.2.0
python-multipart==0.0.6
motor==3.3.2